# products/brain/apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageCreate": {
        "additionalProperties": false,
        "properties": {
          "access_policy_id": {
            "format": "uuid",
            "title": "Access Policy Id",
            "type": "string"
          },
          "content_markdown": {
            "title": "Content Markdown",
            "type": "string"
          },
          "folder_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "position": {
            "default": 0,
            "title": "Position",
            "type": "integer"
          },
          "slug": {
            "maxLength": 63,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceInput"
            },
            "title": "Sources",
            "type": "array"
          },
          "steward_id": {
            "format": "uuid",
            "title": "Steward Id",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "content_markdown",
          "access_policy_id",
          "steward_id"
        ],
        "title": "PageCreate",
        "type": "object"
      },
      "PageInventoryItem": {
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "current_version_id": {
            "format": "uuid",
            "title": "Current Version Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "path",
          "current_version_id",
          "content_hash"
        ],
        "title": "PageInventoryItem",
        "type": "object"
      },
      "PageResponse": {
        "properties": {
          "access_policy_id": {
            "format": "uuid",
            "title": "Access Policy Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "$ref": "#/components/schemas/PageVersionResponse"
          },
          "folder_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "steward_id": {
            "format": "uuid",
            "title": "Steward Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/PageVersionResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "organization_id",
          "folder_id",
          "slug",
          "title",
          "access_policy_id",
          "position",
          "steward_id",
          "created_at",
          "updated_at",
          "current_version",
          "versions"
        ],
        "title": "PageResponse",
        "type": "object"
      },
...
      "PageVersionResponse": {
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "content_markdown": {
            "title": "Content Markdown",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_id": {
            "format": "uuid",
            "title": "Created By Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_id": {
            "format": "uuid",
            "title": "Page Id",
            "type": "string"
          },
          "provenance": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceResponse"
            },
            "title": "Provenance",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "page_id",
          "version",
          "content_markdown",
          "content_hash",
          "created_by_id",
          "created_at",
          "provenance"
        ],
        "title": "PageVersionResponse",
        "type": "object"
      },
      "ProvenanceInput": {
        "additionalProperties": false,
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "relationship": {
            "minLength": 1,
            "title": "Relationship",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": ["source_id", "relationship"],
        "title": "ProvenanceInput",
        "type": "object"
      },
      "ProvenanceResponse": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "relationship": {
            "title": "Relationship",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceResponse"
          }
        },
        "required": ["source", "relationship", "metadata"],
        "title": "ProvenanceResponse",
        "type": "object"
      },
...
      "SourceResponse": {
        "properties": {
          "access_policy_id": {
            "format": "uuid",
            "title": "Access Policy Id",
            "type": "string"
          },
          "canonical_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Uri"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          },
          "provenance": {
            "additionalProperties": true,
            "title": "Provenance",
            "type": "object"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "steward_id": {
            "format": "uuid",
            "title": "Steward Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "source_type",
          "title",
          "canonical_uri",
          "external_id",
          "status",
          "access_policy_id",
          "steward_id",
          "metadata",
          "provenance",
          "created_at",
          "updated_at"
        ],
        "title": "SourceResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pages": {
      "get": {
        "operationId": "list_pages_pages_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PageInventoryItem"
                  },
                  "title": "Response List Pages Pages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pages"
      },
      "post": {
        "operationId": "create_page_pages_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Page"
      }
    },
...
  }
}